requirements.txt
bleach==5.0.1
build==0.9.0
certifi==2022.12.7
charset-normalizer==2.1.1
commonmark==0.9.1
cytoolz==0.12.0
docutils==0.19
eth-hash==0.5.1
eth-typing==3.2.0
eth-utils==2.1.0
idna==3.4
importlib-metadata==5.2.0
jaraco.classes==3.2.3
keyring==23.13.1
more-itertools==9.0.0
nose==1.3.7
packaging==22.0
pep517==0.13.0
pkginfo==1.9.2
plyvel==1.5.0
pre-commit
pycryptodome==3.8.1
Pygments==2.13.0
pysha3==1.0.2
readme-renderer==37.3
requests==2.28.1
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==12.6.0
rlp==3.0.0
six==1.16.0
tomli==2.0.1
toolz==0.9.0
twine==4.0.2
urllib3==1.26.13
webencodings==0.5.1
zipp==3.11.0
